Moved Footer to Commons dir as done for header
Also upped the header one dir up since it was the only file there
