+++ frontend/src/App.tsx
@@ -1,6 +1,7 @@
 import '@bcgov/design-tokens/css/variables.css';
 import '@bcgov/bc-sans/css/BC_Sans.css';
-import Header from './Commons/Header/Header';
+import Header from './Commons/Header';
+import Footer from './Commons/Footer';
 
 const App = () => (
   <>
